+++ src/main/java/lotto/domain/LottoVendingMachine.java
@@ -1,26 +1,7 @@
 package lotto.domain;
 
-
 import java.util.List;
 
-public class LottoVendingMachine {
-
-    public List<Lotto> generateLottos(Integer money) {
-        validate(money);
-        return Lotto.generateByAmount(calculateTotalAmount(money));
-    }
-
-    private void validate(Integer money) {
-        validateIsPositiveMultipleOf1000(money);
-    }
-
-    private void validateIsPositiveMultipleOf1000(Integer money) {
-        if (money < 1000 || money % 1000 != 0) {
-            throw new IllegalArgumentException();
-        }
-
-    }
-    private Integer calculateTotalAmount(Integer money) {
-        return money / 1000;
-    }
+public interface LottoVendingMachine {
+    List<Lotto> generateLottos(Integer money);
 }
