+++ src/test/java/lotto/domain/LottoVendingMachineTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.AppConfig;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -9,7 +10,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 @DisplayName("1000의 양의 배수만큼 돈을 지불하지 않으면 예외가 발생한다")
 class LottoVendingMachineTest {
 
-    LottoVendingMachine lottoVendingMachine = new LottoVendingMachine();
+    AppConfig appConfig = new AppConfig();
+    LottoVendingMachine lottoVendingMachine = appConfig.createLottoVendingMachine();
 
     @DisplayName("정상 입력")
     @Test
